Debug row and col calculation for python
For issue #28, was able to remake the list of arrays that match MATLAB
fdct_wrapping_param output. Need to identify how the indexing works though,
as MATLAB is indexing using X_rows{s}{w}(test), where test is an exorbitantly
high number. Need to find a way to extract the specific index in python without
using these indexing method as it does not exist in Python